basic client

+++ pkg/gochia/Harvester.go
@@ -1,14 +1,14 @@
 package gochia
 
 type Plot struct {
-	FileSize               uint64 `json:"file_size"`
-	Filename               string `json:"filename"`
-	PlotSeed               string `json:"plot-seed"`
-	PlotPublicKey          string `json:"plot_public_key"`
-	PoolContractPuzzleHash string `json:"pool_contract_puzzle_hash"`
-	PoolPublicKey          string `json:"pool_public_key"`
-	Size                   uint64 `json:"size"`
-	TimeModified           uint64 `json:"time_modified"`
+	FileSize               uint64  `json:"file_size"`
+	Filename               string  `json:"filename"`
+	PlotSeed               string  `json:"plot-seed"`
+	PlotPublicKey          string  `json:"plot_public_key"`
+	PoolContractPuzzleHash string  `json:"pool_contract_puzzle_hash"`
+	PoolPublicKey          string  `json:"pool_public_key"`
+	Size                   uint64  `json:"size"`
+	TimeModified           float64 `json:"time_modified"`
 }
 
 type PlotsResponse struct {
